+++ cypress/integration/nav_create_signage.js
@@ -32,7 +32,9 @@ describe('Create signage', () => {
     cy.get("a.pointtopology-control").click()
     cy.get('.leaflet-map-pane')
       .click(403, 287);
-    cy.get("input[name='name']").type('Signage number 1')
+    cy.get("input[name='name_en']").type('Signage number 1')
+    cy.get("a[href='#name_fr']").click()
+    cy.get("input[name='name_fr']").type('Signalétique numéro 1')
     cy.get("select[id='id_type']").select("Service")
     cy.get('#save_changes').click()
     cy.url().should('not.include', '/signage/add/')