+++ geotrek/feedback/models.py
@@ -23,6 +23,8 @@ from mapentity.models import MapEntityMixin
 from geotrek.common.mixins import (AddPropertyMixin, NoDeleteMixin,
                                    PicturesMixin, TimeStampedModelMixin)
 from geotrek.common.utils import intersecting
+from geotrek.core.models import Path
+from geotrek.maintenance.models import Intervention
 from geotrek.trekking.models import POI, Service, Trek
 
 from .helpers import SuricateMessenger
@@ -400,7 +402,8 @@ class Report(MapEntityMixin, PicturesMixin, TimeStampedModelMixin, NoDeleteMixin
             return reports.latest('date_update').get_date_update()
         return cls.objects.none()
 
-Report.add_property('treks', lambda self: intersecting(Path, self), _("Paths"))
+
+Report.add_property('paths', lambda self: intersecting(Path, self), _("Paths"))
 Report.add_property('treks', lambda self: intersecting(Trek, self), _("Treks"))
 Report.add_property('pois', lambda self: intersecting(POI, self), _("POIs"))
 Report.add_property('services', lambda self: intersecting(Service, self), _("Services"))